cache.c: reduce MOVE_SIZE 32->16
The main improvement in the hit ratio comes from moving the insertion
point closer to the end of array.  (I observed the same phenomenon in
rpm-4.0/lib/set.c, where the insertion point was once set to 243/256).

The hit case (as opposed to the miss/insertion case) is somewhat less
important and somewhat more complicated.  The best hit ratio is achieved
when MOVE_SIZE is around 20 or 24.  But the curve is not perfectly
concave, and it turns out the hit ratio is even better with MOVE_SIZE
as low as 8.  (It says, "Don't bother moving anything, all the good
versions are already in the cache".  It's also called "overfitting".)

Another realization is that we shouldn't ultimately optimize for the
best hit ratio.  The ultimate goal is to reduce cycles per rpmsetcmp()
call.  A small drop in the hit ratio can be compensated with faster
cache_move (because there are fewer pointers to move).

Haswell:
-rpmsetcmp: 6700 cycles per call (MOVE_SIZE 32)
+rpmsetcmp: 6460 cycles per call (MOVE_SIZE 16)

On A57, the difference is similar but less pronounced, probably because
of the overhead of the two memmove() calls.

It turns out that moving 64-bit pointers around is somewhat expensive.
If the cache size is limited to 256, then by adding a level of
indirection we can move bytes instead of pointers.

    struct cache {
	uint16_t hv[256];
	uint8_t jv[256];
	struct ent *ev[256];
    };

    // Find by hash.
    hp = cache_find16(hp, h);
    i = hp - C->hv;
    j = C->jv[i];
    e = C->ev[j];

I tried exactly this and it worked faster.  However, here comes the
peril of synthetic benchmarking: it only works faster because, when
calling rpmsetcmp() in a tight loop, ev[] remains in the L1/L2 cache.
If ev[i] is a miss to main memory, the penalty can be much higher
than the benefit.