high half bitstream register pending to be loaded on 32-bit arch
This is one possible way to handle wide bitstream loads on 32-bit
architectures.  If we need to load 10 base64 bytes, we load only 5
bytes, and pointer to high 5 bytes is saved in a variable.  When the
register is exhausted, the next load first checks the saved variable.

The only (relatively minor) problem with this is that it complicates
the loop control: we need to know true bfill, for which we need to
recheck the saved variable.

5-byte bitstream loads vs 10-byte bitstream loads:

$ zstd -d <conv+2.setcmp.zst |awk 'length($1)>=1024{print$1,$1}' |./bench-unmet-sse4
 (x86_64)
-7.55 cycles per integer
+6.85 cycles per integer
 (i686)
-8.38 cycles per integer
+9.04 cycles per integer
 (power9)
-1.76 cycles per integer
+1.63 cycles per integer

The performance drop is i686 is relatively minor: the extra complexity
kicks on only when the register gets reloaded.  Furthermore, the test
for the saved variable being NULL is amenable to branch prediction: it
alternates between NULL and non-NULL.