setstring-dec.c: preload the high half on 32-bit arch
Pre-decoding the data into "int32_t bhi" rather than saving the address
in "const char *bhi" works slightly faster in practice.  The only
problem is that bhi must also be a reliable predicate (because there
is a possibility that the result of decoding is 0, the corresponding
sequence is "AAAAA").  Since the decoded result is a multiple of 6 bits,
there is always a spare high bit which we can flip.  In fact, the
bitwise NOT serves the purpose perfectly.

(i686)
$ zstd -d <conv+2.setcmp.zst |awk 'length($1)>=1024{print$1,$1}' |./bench-unmet-sse4
-9.04 cycles per integer
+8.84 cycles per integer