setstring-dec.c: try to SIMD-load 60 bits
If we do wide loads, why not use SIMD base64 decoding.  We need to
decode 10 bytes, so the result of decoding extra 6 bytes has to be
thrown away.  This actually runs slower than the scalar version.

-6.87 cycles per integer (#if 0)
+6.96 cycles per integer

It seems unlikely that we can gain much from SIMD base64 decoding for q,
as opposed to r.  The difference is that for r, we need m=10 or m=12
bits per value, and only 2 bits (on average) for q.  It's a relatively
small part of the job, so it's harder to capitalize on.