try only Fill(2) and Fill(1) if the remaining len < kq
When the number of remaining characters gets less than 5 or 10, there
is no point in trying to load every possible length between 1 and 10.
Two characters are just fine for tail loads.

It does not make much difference for now (except reducing the code
size); in fact, the output of setstring-enc.c is the same (because
len < kq implies len < kc in the loop control for the next iteration).
This may change as I explore the "fast tail" loop.