setstring-enc.c: rewrite without the block queue, yay!
I found it hard to reason about the encoder that differs from the
decoder in a clever way: len gets decremented immediately, but the block
is pushed to the queue - "not yet written, but must be accounted for".
Another difference is that we have two variables: len and dlen, but dlen
is used only in the loop control, while bitstream flushes seem to rely
on len.  As an exercise, I tried to implement the scheme that explicitly
differentiates between big flushes (kq characters) when the remaining
length is >= 20, and otherwise small flushes (2 characters).  I couldn't
make it work.  While trying to make it work, I resorted to dumping the
internal state on each iteration, because it was not at all clear how
and why the encoder loop might differ from the decoder loop.

Then came an insight from the Higher Intelligence, which reads: there is
only O(1) state associated with the loop.  Thus there is no need for a
push/pop data structure, all the relevant pieces can be held in just a
few local/scalar variables.  Alexey can greatly simplify the loop by
assuming the following invariant: on each iteration, a block gets
written to its final place, specifically at s (much like in the decoder).
After the block is written, Alexey can reserve space for its
complementary q-bits.  There is no need to reserve space more than once
(at two different locations, i.e. there is no second pending location).
This is easier than maintaining the queue, to which we may need to push
more than one, but still O(1), blocks.