setcmploop-x86_64.S: step=5 instead of step=4
We enter the inner loop when we know that *Pv < Rval, to advance Pv.
If Pv[4] >= Rval, we load [1,2,3,4] into Px and do pcmpgtd.  If neither
of [1,2,3] >= Rval, we know that at least [4] >= Rval, so the resulting
mask in %eax is always non-zero, with the lowest bit set indicating the
first element that is >= Rval.

However, we can compare Pv[5] >= Rval, and then still load [1,2,3,4].
If all [1,2,3,4] < Rval, the mask in %eax will be zero, but we can amend
the mask to indicate that at least [5] >= Rval.

-2330 cycles per setcmploop call
+2315 cycles per setcmploop call