setcmploop-x86_64.S: step=8 instead of step=9
-2270 cycles per setcmploop call
+2280 cycles per setcmploop call

The trick with amending the mask still makes sense, although it costs an
extra cycle.  It won't make sense when we load more values than we need
to skip.  We load 8 values, but the L2(cmp) loop is executed only for
Pn >= 32 * Rn, so we need skip 32+ vales on average.