setcmploop-x86_64.S: re-add step=4 loop with cmovs for bpp=32
To use step=9 loop, it is enough to check that Rv[n-1] <= INT32_MAX.
There is no need to check that Pv[n-1] <= INT32_MAX.  Let's see what
happens when we match Rval=Rv[n-1] against the rest of Pv[] which
looks negative to pcmpgtd.

    Pv[0]             Pend                                Pv[9]
     |                 |                                   |
     v                 v                                   v
     +---+ +---+ +---+ +---+ +---+ +---+ +---+ +---+ +---+ +---+
     |   | |NEG| |NEG| |MAX| |MAX| |MAX| |MAX| |MAX| |MAX| |MAX|
     +---+ +---+ +---+ +---+ +---+ +---+ +---+ +---+ +---+ +---+

           |<------------------->| |<------------------->|
                              SIMD scan

If the comparisons were unsigned, this SIMD scan would take us to Pv[1].
With signed comparisons though, the remaining elements look negative
(less than Rval), so the scan takes us all the way to Pv[9] (the outer
sentinel).  After the scan, we can't tell if there were extra elements
in Pv[] that we skipped because they looked negative.  Recall that if
there are extra elements in P that are not in R, we should normally
clear the Le flag.  But this is exactly what we do in L(found9), because
the condition to enter L(found9) is (*Pv < Rval) - there is an element
in P which is not in R, and we want to skip more such elements.

I can't think of a counterexample that would show that it's not enough
to test Rv[n-1] <= INT32_MAX.