cache.c: in hash16, mix in len via the multiplier
This should reduce the latency, because there is no extra step after
multiplication.  (Of course, there are caveats: we now multiply the data
by the data, which can be risky, e.g. we don't want the multiplier to
become zero.  But it works quite well in practice.)

I think I saw this technique in FarmHash.  The constant is from xxHash.
This particular constant seems to work noticeably better than some other
constants.