setcmploop-x86_64.S: after cmp $0, use jl instead of js
cmp+js do not fuse, but cmp+jl do, according to
https://en.wikichip.org/wiki/macro-operation_fusion#x86

Furthermore, fusion should work even when cmp has a memory operand:

    cmp eax, [mem]
    jne loop		    =>	    cmpjne eax, [mem], loop

https://en.wikichip.org/wiki/intel/microarchitectures/sandy_bridge_(client)#Front-end

But our case is even worse, because we compare the memory operand
against a constant, and for fusion to work, there must be enough bits
somewhere to store the constant.  So I'm not sure whether fusion works
with 4-byte constants, but I hope at least it works with $0 (which gets
encoded with a single byte).